+++ packages/components/src/Form/Inputs/InlineInputText/InlineInputText.test.tsx
@@ -28,6 +28,14 @@ import { assertSnapshot } from '@looker/components-test-utils'
 import React from 'react'
 import { InlineInputText } from './InlineInputText'
 
-test('InlineInputText renders an input with the correct styling', () => {
-  assertSnapshot(<InlineInputText title="type something" />)
+test('InlineInputText renders an input with a placeholder', () => {
+  assertSnapshot(<InlineInputText placeholder="this is the placeholder" />)
+})
+
+test('InlineInputText renders an input with no value', () => {
+  assertSnapshot(<InlineInputText />)
+})
+
+test('InlineInputText renders an input with specific predefined value', () => {
+  assertSnapshot(<InlineInputText value="type here..." />)
 })