+++ packages/playground/src/index.tsx
@@ -27,12 +27,12 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import { ComponentsProvider } from '@looker/components'
-import { ActionListDemo } from './ActionList/ActionListDemo'
+import { InlineInputTextDemo } from './InlineInputText/InlineInputTextDemo'
 
 const App: React.FC = () => {
   return (
     <ComponentsProvider>
-      <ActionListDemo />
+      <InlineInputTextDemo />
     </ComponentsProvider>
   )
 }